Set C binding for unix_bind_job in its own C file
